+++ notificacion-service/src/main/java/com/pokedex/notificacion_service/Repository/NotificacionRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface NotificacionRepository extends JpaRepository<Notificacion, Long> {
-    List<Notificacion> findByUserIdOrderByCreatedAtDesc(Integer userId);
+    List<Notificacion> findByUserIdOrderByCreatedAtDesc(Long userId);
 }
